feat: Implement Gemini API integration for project prompt generation
- Added `gemini_api.py` to handle API calls to the Gemini model for generating project documentation.
- Introduced logging for API requests and responses, including error handling and debugging options.
- Created `logger_config.py` for centralized logging configuration.
- Developed `project_generator.py` to manage project analysis, file structure generation, and AI interaction for identifying important files.
- Implemented token calculation and management in `token_utils.py` to ensure compliance with API limits.
- Enhanced file loading logic to respect token limits and improve efficiency.
- Added fallback mechanisms for AI interactions to ensure robustness in file selection and documentation generation.
